pyproject.toml
[project]
name = "deepshika"
version = "0.1.0"
description = "Self-modifying agent experiment — local only, human-in-the-loop"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "ddgs>=8.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "youtube-transcript-api>=0.6",
    "praw>=7.0",
    "pytrends>=4.9",
    "newsapi-python>=0.2",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "google-auth>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
deepshika = "agent:cli"

[tool.setuptools]
py-modules = ["agent", "costs", "tools", "supervisor", "dashboard", "llm", "templates"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
